Fix manifold test
