+++ scripts/bet_omen.py
@@ -1,12 +1,12 @@
 import typer
 
 from prediction_market_agent_tooling.gtypes import PrivateKey, xdai_type
-from prediction_market_agent_tooling.markets.omen.api import (
+from prediction_market_agent_tooling.markets.omen.omen import (
+    OmenAgentMarket,
     get_market,
     omen_buy_outcome_tx,
     omen_sell_outcome_tx,
 )
-from prediction_market_agent_tooling.markets.omen.omen import OmenAgentMarket
 from prediction_market_agent_tooling.tools.web3_utils import verify_address
 
 app = typer.Typer()
